Optimizations for JSON data output

+++ src/main/java/com/jonathanpotts/cartographytable/models/ChunkModel.java
@@ -3,5 +3,5 @@ package com.jonathanpotts.cartographytable.models;
 import java.util.Map;
 
 public class ChunkModel {
-    public Map<Integer, Map<Integer, Map<Integer, BlockModel>>> blocks;
+    public Map<Short, Map<Byte, Map<Byte, BlockModel>>> blocks;
 }
